+++ app/gui2/src/components/GraphEditor/GraphNodes.vue
@@ -1,17 +1,11 @@
 <script setup lang="ts">
 import GraphNode from '@/components/GraphEditor/GraphNode.vue'
-import { SnapGrid, useDragging } from '@/components/GraphEditor/snapGrid'
+import { useDragging } from '@/components/GraphEditor/dragging'
 import { injectGraphNavigator } from '@/providers/graphNavigator'
 import { injectGraphSelection } from '@/providers/graphSelection'
 import { useGraphStore } from '@/stores/graph'
-import { useApproach } from '@/util/animation'
-import { Rect } from '@/util/rect'
-import type { SelectionComposable } from '@/util/selection'
 import { Vec2 } from '@/util/vec2'
-import { iteratorFilter } from 'lib0/iterator'
-import { abs } from 'lib0/math'
 import type { ContentRange, ExprId } from 'shared/yjsModel'
-import { ref, watchEffect, type WatchStopHandle } from 'vue'
 
 const graphStore = useGraphStore()
 const dragging = useDragging()
@@ -28,7 +22,7 @@ function updateNodeContent(id: ExprId, updates: [ContentRange, string][]) {
 
 function nodeIsDragged(movedId: ExprId, offset: Vec2) {
   const scaledOffset = offset.scale(1 / (navigator?.scale ?? 1))
-  dragging.update(movedId, scaledOffset)
+  dragging.startOrUpdate(movedId, scaledOffset)
 }
 
 function hoverNode(id: ExprId | undefined) {
