+++ app/gui2/src/components/GraphEditor/dragging.ts
@@ -4,7 +4,6 @@ import { useApproach } from '@/util/animation'
 import { partitionPoint } from '@/util/array'
 import type { Opt } from '@/util/opt'
 import { Rect } from '@/util/rect'
-import type { SelectionComposable } from '@/util/selection'
 import { Vec2 } from '@/util/vec2'
 import { iteratorFilter } from 'lib0/iterator'
 import { abs } from 'lib0/math'
